fix: normalize line endings during VFS::set_file_contents
