+++ crates/load-cargo/src/lib.rs
@@ -542,7 +542,7 @@ fn load_crate_graph_into_db(
     }
     let changes = vfs.take_changes();
     for (_, file) in changes {
-        if let vfs::Change::Create(v, _) | vfs::Change::Modify(v, _) = file.change
+        if let vfs::Change::Create(v, _, _) | vfs::Change::Modify(v, _, _) = file.change
             && let Ok(text) = String::from_utf8(v)
         {
             analysis_change.change_file(file.file_id, Some(text))
