+++ crates/rust-analyzer/Cargo.toml
@@ -50,7 +50,6 @@ triomphe.workspace = true
 toml.workspace = true
 walkdir = "2.5.0"
 semver.workspace = true
-memchr = "2.7.5"
 cargo_metadata.workspace = true
 process-wrap.workspace = true
 dhat = { version = "0.3.3", optional = true }
